requirements.txt
pandas>=2.0.0
numpy>=1.27.0
scikit-learn>=1.3.0
unittest2>=1.1.0
xgboost>=1.7.5
lightgbm>=4.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
ydata-profiling==4.18.0
optuna>=4.6.0
